+++ packages/embedded-client/src/interval.ts
@@ -1,0 +1,2 @@
+export const defaultInterval = 60 as number;
+export const minimumInterval = 10 as number;