+++ packages/embedded/src/client.ts
@@ -63,11 +63,16 @@ export class Embedded extends Client {
         ? loaderOrSource
         : new Loader(loaderOrSource, options);
 
-    super(loader._transport, {
-      headers: loader._options.headers,
-      userAgent: loader._userAgent,
+    super(loader["~transport"], {
+      headers: loader["~options"].headers,
+      userAgent: loader["~userAgent"],
     });
 
     this.loader = loader;
+  }
+
+  /** @internal */
+  public override get ["~updateSignal"](): unknown {
+    return this.loader["~active"];
   }
 }