+++ packages/embedded/src/loader.ts
@@ -1,11 +1,12 @@
 import type {
+  Awaitable,
   Options as CoreOptions,
   DecisionLogEntry,
   JWT,
   SourceAttributes,
   Value,
 } from "@cerbos/core";
-import { _setErrorNameAndStack } from "@cerbos/core";
+import { setErrorNameAndStack, userAgent } from "@cerbos/core/~internal";
 
 import { Bundle, download } from "./bundle";
 import { constrainAutoUpdateInterval } from "./interval";
@@ -27,9 +28,7 @@ type LoadResult =
       error: LoadError;
     };
 
-async function resolve(
-  result: LoadResult | Promise<LoadResult>,
-): Promise<Bundle> {
+async function resolve(result: Awaitable<LoadResult>): Promise<Bundle> {
   const { bundle, error } = await result;
 
   if (error) {
@@ -57,7 +56,7 @@ export class LoadError extends Error {
       cause,
     });
 
-    _setErrorNameAndStack(this);
+    setErrorNameAndStack(this);
   }
 }
 
@@ -69,13 +68,10 @@ export class LoadError extends Error {
 export type Source =
   | string
   | URL
-  | ArrayBufferView<ArrayBuffer>
-  | ArrayBuffer
-  | Response
-  | WebAssembly.Module
-  | Promise<
-      ArrayBufferView<ArrayBuffer> | ArrayBuffer | Response | WebAssembly.Module
-    >;
+  | Awaitable<ArrayBufferView<ArrayBuffer>>
+  | Awaitable<ArrayBuffer>
+  | Awaitable<Response>
+  | Awaitable<WebAssembly.Module>;
 
 /**
  * Options for creating a new {@link Embedded} client or {@link Loader}.
@@ -130,21 +126,21 @@ export interface Options extends Pick<CoreOptions, "headers" | "userAgent"> {
    *
    * @defaultValue (no-op)
    */
-  onLoad?: ((metadata: BundleMetadata) => void | Promise<void>) | undefined;
+  onLoad?: ((metadata: BundleMetadata) => Awaitable<void>) | undefined;
 
   /**
    * A callback to invoke when the embedded policy decision point bundle has failed to load.
    *
    * @defaultValue (no-op)
    */
-  onError?: ((error: LoadError) => void | Promise<void>) | undefined;
+  onError?: ((error: LoadError) => Awaitable<void>) | undefined;
 
   /**
    * A callback to invoke when a decision is made by the embedded policy decision point.
    *
    * @defaultValue (no-op)
    */
-  onDecision?: ((entry: DecisionLogEntry) => void | Promise<void>) | undefined;
+  onDecision?: ((entry: DecisionLogEntry) => Awaitable<void>) | undefined;
 }
 
 /**
@@ -201,9 +197,7 @@ export interface Options extends Pick<CoreOptions, "headers" | "userAgent"> {
  *
  * @public
  */
-export type DecodeJWTPayload = (
-  jwt: JWT,
-) => DecodedJWTPayload | Promise<DecodedJWTPayload>;
+export type DecodeJWTPayload = (jwt: JWT) => Awaitable<DecodedJWTPayload>;
 
 /**
  * The decoded payload of a JWT, containing the claims.
@@ -253,13 +247,18 @@ export interface BundleMetadata {
  */
 export class Loader {
   /** @internal */
-  public _active: LoadResult | Promise<LoadResult>;
+  public readonly ["~options"]: Options;
 
   /** @internal */
-  public readonly _options: Options;
+  public readonly ["~userAgent"]: string;
 
   /** @internal */
-  public readonly _userAgent: string;
+  public readonly ["~transport"] = new Transport(
+    async () => await resolve(this["~active"]),
+  );
+
+  /** @internal */
+  protected ["~active"]: Awaitable<LoadResult>;
 
   private readonly logger: DecisionLogger | undefined;
 
@@ -302,33 +301,30 @@ export class Loader {
    * ```
    */
   public constructor(source: Source, options: Options = {}) {
-    this._options = options;
-
-    this._userAgent = `${
-      options.userAgent ? `${options.userAgent} ` : ""
-    }${defaultUserAgent}`;
+    this["~options"] = options;
+    this["~userAgent"] = userAgent(options.userAgent, defaultUserAgent);
 
     if (options.onDecision) {
-      this.logger = new DecisionLogger(options.onDecision, this._userAgent);
+      this.logger = new DecisionLogger(options.onDecision, this["~userAgent"]);
     }
 
-    this._active = this._load(source, url(source), true);
+    this["~active"] = this["~load"](source, url(source), true);
   }
 
   /**
    * Resolves to the metadata of the loaded bundle, or rejects with the error that was encountered when loading the bundle.
    */
   public async active(): Promise<BundleMetadata> {
-    return (await resolve(this._active)).metadata;
+    return (await resolve(this["~active"])).metadata;
   }
 
   /** @internal */
-  public readonly _transport = new Transport(
-    async () => await resolve(this._active),
-  );
+  public get ["~updateSignal"](): unknown {
+    return undefined;
+  }
 
   /** @internal */
-  protected async _load(
+  protected async ["~load"](
     source: Source,
     url: string | undefined,
     initial = false,
@@ -338,28 +334,31 @@ export class Loader {
         source,
         url,
         this.logger,
-        this._userAgent,
-        this._options,
+        this["~userAgent"],
+        this["~options"],
       );
 
-      await this._onLoad(bundle, initial);
+      await this["~onLoad"](bundle, initial);
 
       return { bundle };
     } catch (cause) {
       const error = new LoadError(cause);
-      await this._onError(error);
+      await this["~onError"](error);
       return { error };
     }
   }
 
   /** @internal */
-  protected async _onLoad(bundle: Bundle, _initial: boolean): Promise<void> {
-    await this._options.onLoad?.(bundle.metadata);
+  protected async ["~onLoad"](
+    bundle: Bundle,
+    _initial: boolean,
+  ): Promise<void> {
+    await this["~options"].onLoad?.(bundle.metadata);
   }
 
   /** @internal */
-  protected async _onError(error: LoadError): Promise<void> {
-    await this._options.onError?.(error);
+  protected async ["~onError"](error: LoadError): Promise<void> {
+    await this["~options"].onError?.(error);
   }
 }
 
@@ -403,6 +402,7 @@ export interface AutoUpdateOptions extends Options {
 export class AutoUpdatingLoader extends Loader {
   private readonly activateOnLoad: boolean;
   private readonly interval: number;
+  private activations = 0;
   private _pending: Bundle | undefined;
   private etag: string | undefined;
   private running = true;
@@ -455,7 +455,8 @@ export class AutoUpdatingLoader extends Loader {
    */
   public activate(): void {
     if (this._pending) {
-      this._active = { bundle: this._pending };
+      this.activations++;
+      this["~active"] = { bundle: this._pending };
       this._pending = undefined;
     }
   }
@@ -473,7 +474,12 @@ export class AutoUpdatingLoader extends Loader {
   }
 
   /** @internal */
-  protected override async _onLoad(
+  public override get ["~updateSignal"](): unknown {
+    return this.activations;
+  }
+
+  /** @internal */
+  protected override async ["~onLoad"](
     bundle: Bundle,
     initial: boolean,
   ): Promise<void> {
@@ -487,13 +493,13 @@ export class AutoUpdatingLoader extends Loader {
       }
     }
 
-    await super._onLoad(bundle, initial);
+    await super["~onLoad"](bundle, initial);
   }
 
   /** @internal */
-  protected override async _onError(error: LoadError): Promise<void> {
+  protected override async ["~onError"](error: LoadError): Promise<void> {
     if (!this.suppressError(error.cause)) {
-      await super._onError(error);
+      await super["~onError"](error);
     }
   }
 
@@ -515,7 +521,7 @@ export class AutoUpdatingLoader extends Loader {
     this.abortController?.abort();
     this.abortController = new AbortController();
 
-    await this._load(
+    await this["~load"](
       this.download(this.abortController.signal),
       this.url.toString(),
     );
@@ -530,7 +536,7 @@ export class AutoUpdatingLoader extends Loader {
       request.headers = { "If-None-Match": this.etag };
     }
 
-    const response = await download(this.url, this._userAgent, request);
+    const response = await download(this.url, this["~userAgent"], request);
 
     if (response.status === 304) {
       cancelBody(response);