+++ private/test/src/setup.ts
@@ -2,7 +2,8 @@ import "@ungap/with-resolvers";
 
 import { Module } from "module";
 
-import type * as interval from "../../../packages/embedded/src/interval";
+import type * as embeddedV1Interval from "../../../packages/embedded/src/interval";
+import type * as embeddedV2Interval from "../../../packages/embedded-client/src/interval";
 
 const loader = Module as unknown as {
   _load: (request: string, parent: { id: string }, isMain: boolean) => unknown;
@@ -17,7 +18,17 @@ loader._load = (request, parent, isMain): unknown => {
   ) {
     return {
       constrainAutoUpdateInterval: (interval) => interval ?? 100,
-    } satisfies typeof interval;
+    } satisfies typeof embeddedV1Interval;
+  }
+
+  if (
+    request === "./interval" &&
+    parent.id.endsWith("/packages/embedded-client/lib/loader.js")
+  ) {
+    return {
+      defaultInterval: 0.1,
+      minimumInterval: 0.1,
+    } satisfies typeof embeddedV2Interval;
   }
 
   return load(request, parent, isMain);